requirements.txt
aiogram==2.25.2
uvloop==0.19.0
ujson==5.9.0
aiohttp[speedups]
redis

# Other packages
python-dotenv
pydantic
